+++ src/BayesFilters/include/BayesFilters/LinearModel.h
@@ -84,7 +84,7 @@ protected:
      */
     std::function<double()> gauss_rnd_sample_;
 
-    std::vector<std::string> log_filenames(const std::string& prefix_path, const std::string& prefix_name) override
+    std::vector<std::string> log_file_names(const std::string& prefix_path, const std::string& prefix_name) override
     {
         return {prefix_path + "/" + prefix_name + "_measurements"};
     }