+++ src/BayesFilters/include/BayesFilters/Logger.h
@@ -53,7 +53,7 @@ public:
     }
 
 protected:
-    virtual std::vector<std::string> log_filenames(const std::string& prefix_path, const std::string& prefix_name);
+    virtual std::vector<std::string> log_file_names(const std::string& prefix_path, const std::string& prefix_name);
 
     virtual void log();