+++ src/BayesFilters/include/BayesFilters/SimulatedStateModel.h
@@ -46,7 +46,7 @@ protected:
 
     mutable std::ofstream log_file_state_;
 
-    std::vector<std::string> log_filenames(const std::string& prefix_path, const std::string& prefix_name) override
+    std::vector<std::string> log_file_names(const std::string& prefix_path, const std::string& prefix_name) override
     {
         return {prefix_path + "/" + prefix_name + "_target"};
     }