+++ test/test_mixed_UKF_KF/main.cpp
@@ -40,7 +40,7 @@ protected:
     }
 
 
-    std::vector<std::string> log_filenames(const std::string& prefix_path, const std::string& prefix_name) override
+    std::vector<std::string> log_file_names(const std::string& prefix_path, const std::string& prefix_name) override
     {
         return  {prefix_path + "/" + prefix_name + "_pred_mean",
                  prefix_path + "/" + prefix_name + "_cor_mean"};